+++ NEWS.md
@@ -5,6 +5,7 @@
 *  Added the `eda_fval` function that generates f-values (probabilities) based on one of six algorithms
 *  Added the `eda_theopan` function that generates multi-panel theoretical QQ plots.
 *  Added the `eda_qqpool` function that generates multi-panel pooled residual plots.
+*  Added the `eda_shuffle` function that permutes values across groups.
 *  `eda_sl` now outputs Cleveland's spread-location plot as well as Tukey's spread-versus-level plot.
 *  Added `eda_untie` function that splits ties in a vector by nudging them by a small amount.
 *  Fixed bug in `eda_rline` where incorrect residuals were returned.